feat: add admin role restrictions

+++ app/controllers/chefs_controller.rb
@@ -1,5 +1,7 @@
 class ChefsController < ApplicationController
     before_action :set_chef, only: [:show, :edit, :update, :destroy]
+    before_action :require_same_user, only: [:edit, :update]
+    before_action :require_admin, only: [:destroy]
 
     def index
         @chefs = Chef.paginate(page: params[:page], per_page: 5)
@@ -38,12 +40,14 @@ class ChefsController < ApplicationController
     end
 
     def destroy
-        if @chef.destroy
-            flash[:success] = "Chef deleted successfully"
-            redirect_to chefs_path
-        else
-            flash.now[:danger] = "There was an error deleting the chef"
-            render 'index'
+        unless @chef.admin
+            if @chef.destroy
+                flash[:success] = "Chef deleted successfully"
+                redirect_to chefs_path
+            else
+                flash.now[:danger] = "There was an error deleting the chef"
+                render 'index'
+            end
         end
     end
 
@@ -53,5 +57,17 @@ class ChefsController < ApplicationController
     end
     def chef_params
         params.require(:chef).permit(:chefname, :email, :password, :password_confirmation)
+    end
+    def require_same_user
+        if current_chef != @chef && !current_chef.admin?
+            flash[:danger] = "You can only edit or delete your own account"
+            redirect_to chefs_path
+        end
+    end
+    def require_admin
+        if logged_in? && !current_chef.admin?
+            flash[:danger] = "Only admin users can perform that action"
+            redirect_to root_path
+        end
     end
 end
